+++ src/index.ts
@@ -1,5 +1,5 @@
 import { LitElement, html, css } from "lit";
-import { customElement, property, state, query } from "lit/decorators.js";
+import { customElement, property, query } from "lit/decorators.js";
 import { KanbanController } from "./controllers/kanban";
 
 import "./view/Column";
@@ -29,6 +29,8 @@ export class KanbanBoard extends LitElement {
       display: block;
       width: 100%;
       height: auto;
+      text-wrap: initial;
+      white-space-collapse: initial;
     }
 
     .kanban {
@@ -49,12 +51,49 @@ export class KanbanBoard extends LitElement {
 			border-right: 1px solid rgba(120, 120, 120, 0.9);
 			padding-right: 15px;
     }
+
+    dialog {
+      z-index: 10;
+      margin-top: 10px;
+      background: rgba(120, 120, 120, 0.1);
+      border: none;
+      border-radius: 1rem;
+      margin: auto;
+    }
+
+    dialog::backdrop {
+      background-color: rgba(120, 120, 120, 0.25);
+    }
+
+    dialog button {
+      width: 45%;
+      padding: 10px 5px;
+      margin-top: 10px;
+      font-size: 16px;
+      color: inherit;
+      background: rgba(0, 0, 0, 0.1);
+      border: none;
+      border-radius: 5px;
+      cursor: pointer;
+      font-family: inherit;
+    }
+
+    dialog button:not(:last-child) {
+      margin-right: 15px;
+    }
+
+    dialog button:hover {
+      background: rgba(0, 0, 0, 0.2);
+    }
   `;
 
   @property({ reflect: true, type: Object  })
   data: KanbanBoardData = {
     columns: [],
   };
+
+  @query("dialog")
+  _dialog: HTMLDialogElement;
 
   connectedCallback() {
     super.connectedCallback();
@@ -81,7 +120,21 @@ export class KanbanBoard extends LitElement {
           id="${column.id}"
           title="${column.title}"
           items="${JSON.stringify(column.items)}"
-        ></kanban-column>`;
+        ></kanban-column>
+        <!-- A modal dialog containing a form -->
+        <dialog id="favDialog">
+          <form>
+            <p>
+              Are you sure you want to delete this item?
+            </p>
+            <div>
+              <button value="cancel" formmethod="dialog">Cancel</button>
+              <button id="confirmBtn" value="yes">Confirm</button>
+            </div>
+          </form>
+        </dialog>
+        
+        `;
       })}
     </div>`;
   }
@@ -106,11 +159,43 @@ export class KanbanBoard extends LitElement {
     this.kanbanAPI.updateItem(e.detail.id, { content: e.detail.content });
   }
 
-  private _itemDeleteHandler = (e: CustomEvent) =>  {
-    this.kanbanAPI.deleteItem(e.detail.id);
+  private _itemDeleteHandler = (e: CustomEvent) => {
+    const confirmBtn = this._dialog.querySelector("#confirmBtn") as HTMLButtonElement;
+
+    // "Show the dialog" opens the <dialog> modally
+    this._dialog.showModal();
+
+    this._dialog.addEventListener("click", e => {
+      const dialogDimensions = this._dialog.getBoundingClientRect()
+      if (
+        e.clientX < dialogDimensions.left ||
+        e.clientX > dialogDimensions.right ||
+        e.clientY < dialogDimensions.top ||
+        e.clientY > dialogDimensions.bottom
+      ) {
+        this._dialog.close("cancel")
+      }
+    })
+
+    // Prevent the "confirm" button from the default behavior of submitting the form, and close the dialog with the `close()` method, which triggers the "close" event.
+    confirmBtn.addEventListener("click", (event) => {
+      event.preventDefault(); // We don't want to submit this fake form
+      this._dialog.close(confirmBtn.value); // Have to send the select box value here.
+    });
+
+    // "Cancel" button closes the dialog without submitting because of [formmethod="dialog"], triggering a close event.
+    this._dialog.addEventListener("close", (_event) => {
+      if (this._dialog.returnValue === "yes") {
+        this.kanbanAPI.deleteItem(e.detail.id);
+      }
+    });
   }
 
   private _itemAddHandler = (e: CustomEvent) => {
     this.kanbanAPI.insertItem(e.detail.columnId, e.detail.item);
   }
+
+
+
+
 }
